rename datatype_editors to datatype_uis
