+++ crates/viewer/re_component_ui/src/datatype_uis/enum_combobox.rs
